Bitrot cleanup: Explicitly cast GLICT ints to uint16_t when storing some sizes inside SDL_Rect.

+++ spritegl.cpp
@@ -85,8 +85,8 @@ void SpriteGL::buildGLTexture() {
 
 	SDL_Surface *sfc = SDL_CreateRGBSurface(SDL_SWSURFACE | SDL_SRCALPHA, nextpow(getBasicImage()->w), nextpow(getBasicImage()->h), 32, rmask, gmask, bmask, amask);
 	
-	SDL_Rect s = {0,0,getBasicImage()->w,getBasicImage()->h};
-	SDL_Rect d = {0,0,getBasicImage()->w,getBasicImage()->h};
+	SDL_Rect s = {0,0,uint16_t(getBasicImage()->w),uint16_t(getBasicImage()->h)};
+	SDL_Rect d = {0,0,uint16_t(getBasicImage()->w),uint16_t(getBasicImage()->h)};
 
 	SDL_BlitSurface(getBasicImage(), &s, sfc, &d);
